feat(cu128): Blackwell/cu128 uv stack (SDPA attn, torch-norm, bf16 DiT) + uv project + onboarding
Make SeedVR2-3B run on Blackwell (sm_120) / Ubuntu 20.04 (glibc 2.31, no nvcc):
- torch 2.8.0+cu128 via uv (pyproject.toml + uv.lock, pytorch-cu128 index)
- attention: SDPA drop-in for flash_attn_varlen_func (prebuilt flash-attn needs glibc>=2.32)
- normalization: pure-torch rms/layer instead of apex fused (needs GLIBCXX_3.4.29)
- DiT loaded in bf16 to fit 32GB VRAM
- fix upstream image-input output-write bug in inference_seedvr2_3b.py
- docs/ONBOARDING.md for setup/run; color_fix.py kept out of repo (fetch per docs)

+++ projects/video_diffusion_sr/infer.py
@@ -93,7 +93,9 @@ class VideoDiffusionInfer():
             self.dit = meta_non_persistent_buffer_init_fn(self.dit)
 
         if device in [get_device(), "cuda"]:
-            self.dit.to(get_device())
+            # Cast DiT weights to bf16 on GPU (forward runs under autocast-bf16
+            # anyway) to fit the 3B model on a 32GB card instead of fp32 ~12GB.
+            self.dit.to(get_device(), dtype=torch.bfloat16)
 
         # Print model size.
         num_params = sum(p.numel() for p in self.dit.parameters() if p.requires_grad)
